Module aws_sdk_route53resolver::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AssociateFirewallRuleGroup

Fluent builder constructing a request to AssociateFirewallRuleGroup.

AssociateResolverEndpointIpAddress

Fluent builder constructing a request to AssociateResolverEndpointIpAddress.

AssociateResolverQueryLogConfig

Fluent builder constructing a request to AssociateResolverQueryLogConfig.

AssociateResolverRule

Fluent builder constructing a request to AssociateResolverRule.

CreateFirewallDomainList

Fluent builder constructing a request to CreateFirewallDomainList.

CreateFirewallRule

Fluent builder constructing a request to CreateFirewallRule.

CreateFirewallRuleGroup

Fluent builder constructing a request to CreateFirewallRuleGroup.

CreateResolverEndpoint

Fluent builder constructing a request to CreateResolverEndpoint.

CreateResolverQueryLogConfig

Fluent builder constructing a request to CreateResolverQueryLogConfig.

CreateResolverRule

Fluent builder constructing a request to CreateResolverRule.

DeleteFirewallDomainList

Fluent builder constructing a request to DeleteFirewallDomainList.

DeleteFirewallRule

Fluent builder constructing a request to DeleteFirewallRule.

DeleteFirewallRuleGroup

Fluent builder constructing a request to DeleteFirewallRuleGroup.

DeleteResolverEndpoint

Fluent builder constructing a request to DeleteResolverEndpoint.

DeleteResolverQueryLogConfig

Fluent builder constructing a request to DeleteResolverQueryLogConfig.

DeleteResolverRule

Fluent builder constructing a request to DeleteResolverRule.

DisassociateFirewallRuleGroup

Fluent builder constructing a request to DisassociateFirewallRuleGroup.

DisassociateResolverEndpointIpAddress

Fluent builder constructing a request to DisassociateResolverEndpointIpAddress.

DisassociateResolverQueryLogConfig

Fluent builder constructing a request to DisassociateResolverQueryLogConfig.

DisassociateResolverRule

Fluent builder constructing a request to DisassociateResolverRule.

GetFirewallConfig

Fluent builder constructing a request to GetFirewallConfig.

GetFirewallDomainList

Fluent builder constructing a request to GetFirewallDomainList.

GetFirewallRuleGroup

Fluent builder constructing a request to GetFirewallRuleGroup.

GetFirewallRuleGroupAssociation

Fluent builder constructing a request to GetFirewallRuleGroupAssociation.

GetFirewallRuleGroupPolicy

Fluent builder constructing a request to GetFirewallRuleGroupPolicy.

GetResolverConfig

Fluent builder constructing a request to GetResolverConfig.

GetResolverDnssecConfig

Fluent builder constructing a request to GetResolverDnssecConfig.

GetResolverEndpoint

Fluent builder constructing a request to GetResolverEndpoint.

GetResolverQueryLogConfig

Fluent builder constructing a request to GetResolverQueryLogConfig.

GetResolverQueryLogConfigAssociation

Fluent builder constructing a request to GetResolverQueryLogConfigAssociation.

GetResolverQueryLogConfigPolicy

Fluent builder constructing a request to GetResolverQueryLogConfigPolicy.

GetResolverRule

Fluent builder constructing a request to GetResolverRule.

GetResolverRuleAssociation

Fluent builder constructing a request to GetResolverRuleAssociation.

GetResolverRulePolicy

Fluent builder constructing a request to GetResolverRulePolicy.

ImportFirewallDomains

Fluent builder constructing a request to ImportFirewallDomains.

ListFirewallConfigs

Fluent builder constructing a request to ListFirewallConfigs.

ListFirewallDomainLists

Fluent builder constructing a request to ListFirewallDomainLists.

ListFirewallDomains

Fluent builder constructing a request to ListFirewallDomains.

ListFirewallRuleGroupAssociations

Fluent builder constructing a request to ListFirewallRuleGroupAssociations.

ListFirewallRuleGroups

Fluent builder constructing a request to ListFirewallRuleGroups.

ListFirewallRules

Fluent builder constructing a request to ListFirewallRules.

ListResolverConfigs

Fluent builder constructing a request to ListResolverConfigs.

ListResolverDnssecConfigs

Fluent builder constructing a request to ListResolverDnssecConfigs.

ListResolverEndpointIpAddresses

Fluent builder constructing a request to ListResolverEndpointIpAddresses.

ListResolverEndpoints

Fluent builder constructing a request to ListResolverEndpoints.

ListResolverQueryLogConfigAssociations

Fluent builder constructing a request to ListResolverQueryLogConfigAssociations.

ListResolverQueryLogConfigs

Fluent builder constructing a request to ListResolverQueryLogConfigs.

ListResolverRuleAssociations

Fluent builder constructing a request to ListResolverRuleAssociations.

ListResolverRules

Fluent builder constructing a request to ListResolverRules.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

PutFirewallRuleGroupPolicy

Fluent builder constructing a request to PutFirewallRuleGroupPolicy.

PutResolverQueryLogConfigPolicy

Fluent builder constructing a request to PutResolverQueryLogConfigPolicy.

PutResolverRulePolicy

Fluent builder constructing a request to PutResolverRulePolicy.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateFirewallConfig

Fluent builder constructing a request to UpdateFirewallConfig.

UpdateFirewallDomains

Fluent builder constructing a request to UpdateFirewallDomains.

UpdateFirewallRule

Fluent builder constructing a request to UpdateFirewallRule.

UpdateFirewallRuleGroupAssociation

Fluent builder constructing a request to UpdateFirewallRuleGroupAssociation.

UpdateResolverConfig

Fluent builder constructing a request to UpdateResolverConfig.

UpdateResolverDnssecConfig

Fluent builder constructing a request to UpdateResolverDnssecConfig.

UpdateResolverEndpoint

Fluent builder constructing a request to UpdateResolverEndpoint.

UpdateResolverRule

Fluent builder constructing a request to UpdateResolverRule.