Module aws_sdk_route53resolver::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_service_error_exception

See InternalServiceErrorException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_policy_document

See InvalidPolicyDocument

invalid_request_exception

See InvalidRequestException

invalid_tag_exception

See InvalidTagException

limit_exceeded_exception

See LimitExceededException

resource_exists_exception

See ResourceExistsException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_unavailable_exception

See ResourceUnavailableException

throttling_exception

See ThrottlingException

unknown_resource_exception

See UnknownResourceException

validation_exception

See ValidationException

Structs

AccessDeniedException

The current account doesn't have the IAM permissions required to perform the specified Resolver operation.

AssociateFirewallRuleGroupError

AssociateResolverEndpointIpAddressError

AssociateResolverQueryLogConfigError

AssociateResolverRuleError

ConflictException

CreateFirewallDomainListError

CreateFirewallRuleError

CreateFirewallRuleGroupError

CreateResolverEndpointError

CreateResolverQueryLogConfigError

CreateResolverRuleError

DeleteFirewallDomainListError

DeleteFirewallRuleError

DeleteFirewallRuleGroupError

DeleteResolverEndpointError

DeleteResolverQueryLogConfigError

DeleteResolverRuleError

DisassociateFirewallRuleGroupError

DisassociateResolverEndpointIpAddressError

DisassociateResolverQueryLogConfigError

DisassociateResolverRuleError

GetFirewallConfigError

GetFirewallDomainListError

GetFirewallRuleGroupAssociationError

GetFirewallRuleGroupError

GetFirewallRuleGroupPolicyError

GetResolverDnssecConfigError

GetResolverEndpointError

GetResolverQueryLogConfigAssociationError

GetResolverQueryLogConfigError

GetResolverQueryLogConfigPolicyError

GetResolverRuleAssociationError

GetResolverRuleError

GetResolverRulePolicyError

ImportFirewallDomainsError

InternalServiceErrorException

We encountered an unknown error. Try again in a few minutes.

InvalidNextTokenException

The value that you specified for NextToken in a List request isn't valid.

InvalidParameterException

One or more parameters in this request are not valid.

InvalidPolicyDocument

The specified Resolver rule policy is invalid.

InvalidRequestException

The request is invalid.

InvalidTagException

The specified tag is invalid.

LimitExceededException

The request caused one or more limits to be exceeded.

ListFirewallConfigsError

ListFirewallDomainListsError

ListFirewallDomainsError

ListFirewallRuleGroupAssociationsError

ListFirewallRuleGroupsError

ListFirewallRulesError

ListResolverDnssecConfigsError

ListResolverEndpointIpAddressesError

ListResolverEndpointsError

ListResolverQueryLogConfigAssociationsError

ListResolverQueryLogConfigsError

ListResolverRuleAssociationsError

ListResolverRulesError

ListTagsForResourceError

PutFirewallRuleGroupPolicyError

PutResolverQueryLogConfigPolicyError

PutResolverRulePolicyError

ResourceExistsException

The resource that you tried to create already exists.

ResourceInUseException

The resource that you tried to update or delete is currently in use.

ResourceNotFoundException

The specified resource doesn't exist.

ResourceUnavailableException

The specified resource isn't available.

TagResourceError

ThrottlingException

The request was throttled. Try again in a few minutes.

UnknownResourceException

The specified resource doesn't exist.

UntagResourceError

UpdateFirewallConfigError

UpdateFirewallDomainsError

UpdateFirewallRuleError

UpdateFirewallRuleGroupAssociationError

UpdateResolverDnssecConfigError

UpdateResolverEndpointError

UpdateResolverRuleError

ValidationException

Enums

AssociateFirewallRuleGroupErrorKind

AssociateResolverEndpointIpAddressErrorKind

AssociateResolverQueryLogConfigErrorKind

AssociateResolverRuleErrorKind

CreateFirewallDomainListErrorKind

CreateFirewallRuleErrorKind

CreateFirewallRuleGroupErrorKind

CreateResolverEndpointErrorKind

CreateResolverQueryLogConfigErrorKind

CreateResolverRuleErrorKind

DeleteFirewallDomainListErrorKind

DeleteFirewallRuleErrorKind

DeleteFirewallRuleGroupErrorKind

DeleteResolverEndpointErrorKind

DeleteResolverQueryLogConfigErrorKind

DeleteResolverRuleErrorKind

DisassociateFirewallRuleGroupErrorKind

DisassociateResolverEndpointIpAddressErrorKind

DisassociateResolverQueryLogConfigErrorKind

DisassociateResolverRuleErrorKind

GetFirewallConfigErrorKind

GetFirewallDomainListErrorKind

GetFirewallRuleGroupAssociationErrorKind

GetFirewallRuleGroupErrorKind

GetFirewallRuleGroupPolicyErrorKind

GetResolverDnssecConfigErrorKind

GetResolverEndpointErrorKind

GetResolverQueryLogConfigAssociationErrorKind

GetResolverQueryLogConfigErrorKind

GetResolverQueryLogConfigPolicyErrorKind

GetResolverRuleAssociationErrorKind

GetResolverRuleErrorKind

GetResolverRulePolicyErrorKind

ImportFirewallDomainsErrorKind

ListFirewallConfigsErrorKind

ListFirewallDomainListsErrorKind

ListFirewallDomainsErrorKind

ListFirewallRuleGroupAssociationsErrorKind

ListFirewallRuleGroupsErrorKind

ListFirewallRulesErrorKind

ListResolverDnssecConfigsErrorKind

ListResolverEndpointIpAddressesErrorKind

ListResolverEndpointsErrorKind

ListResolverQueryLogConfigAssociationsErrorKind

ListResolverQueryLogConfigsErrorKind

ListResolverRuleAssociationsErrorKind

ListResolverRulesErrorKind

ListTagsForResourceErrorKind

PutFirewallRuleGroupPolicyErrorKind

PutResolverQueryLogConfigPolicyErrorKind

PutResolverRulePolicyErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFirewallConfigErrorKind

UpdateFirewallDomainsErrorKind

UpdateFirewallRuleErrorKind

UpdateFirewallRuleGroupAssociationErrorKind

UpdateResolverDnssecConfigErrorKind

UpdateResolverEndpointErrorKind

UpdateResolverRuleErrorKind