Module aws_sdk_route53resolver::model

Modules

filter

See Filter

firewall_config

See FirewallConfig

firewall_domain_list

See FirewallDomainList

firewall_domain_list_metadata

See FirewallDomainListMetadata

firewall_rule

See FirewallRule

firewall_rule_group

See FirewallRuleGroup

firewall_rule_group_association

See FirewallRuleGroupAssociation

firewall_rule_group_metadata

See FirewallRuleGroupMetadata

ip_address_request

See IpAddressRequest

ip_address_response

See IpAddressResponse

ip_address_update

See IpAddressUpdate

resolver_dnssec_config

See ResolverDnssecConfig

resolver_endpoint

See ResolverEndpoint

resolver_query_log_config

See ResolverQueryLogConfig

resolver_query_log_config_association

See ResolverQueryLogConfigAssociation

resolver_rule

See ResolverRule

resolver_rule_association

See ResolverRuleAssociation

resolver_rule_config

See ResolverRuleConfig

tag

See Tag

target_address

See TargetAddress

Structs

Filter

For Resolver list operations

FirewallConfig

Configuration of the firewall behavior provided by DNS Firewall for a single VPC from

FirewallDomainList

High-level information about a list of firewall domains for use in a FirewallRule. This is returned by GetFirewallDomainList.

FirewallDomainListMetadata

Minimal high-level information for a firewall domain list. The action ListFirewallDomainLists returns an array of these objects.

FirewallRule

A single firewall rule in a rule group.

FirewallRuleGroup

High-level information for a firewall rule group. A firewall rule group is a collection of rules that DNS Firewall uses to filter DNS network traffic for a VPC. To retrieve the rules for the rule group, call ListFirewallRules.

FirewallRuleGroupAssociation

An association between a firewall rule group and a VPC, which enables DNS filtering for

FirewallRuleGroupMetadata

Minimal high-level information for a firewall rule group. The action ListFirewallRuleGroups returns an array of these objects.

IpAddressRequest

In a

IpAddressResponse

In the response to a

IpAddressUpdate

In an

ResolverDnssecConfig

A complex type that contains information about a configuration for DNSSEC validation.

ResolverEndpoint

In the response to a

ResolverQueryLogConfig

In the response to a

ResolverQueryLogConfigAssociation

In the response to an

ResolverRule

For queries that originate in your VPC, detailed information about a Resolver rule, which specifies how to route DNS queries

ResolverRuleAssociation

In the response to an

ResolverRuleConfig

In an

Tag

One tag that you want to add to the specified resource. A tag consists of a Key (a name for the tag) and a Value.

TargetAddress

In a

Enums

Action

BlockOverrideDnsType

BlockResponse

FirewallDomainImportOperation

FirewallDomainListStatus

FirewallDomainUpdateOperation

FirewallFailOpenStatus

FirewallRuleGroupAssociationStatus

FirewallRuleGroupStatus

IpAddressStatus

MutationProtectionStatus

ResolverDnssecValidationStatus

ResolverEndpointDirection

ResolverEndpointStatus

ResolverQueryLogConfigAssociationError

ResolverQueryLogConfigAssociationStatus

ResolverQueryLogConfigStatus

ResolverRuleAssociationStatus

ResolverRuleStatus

RuleTypeOption

ShareStatus

SortOrder

Validation