Module aws_sdk_route53resolver::operation

Structs

AssociateFirewallRuleGroup

Associates a FirewallRuleGroup with a VPC, to provide DNS filtering for the VPC.

AssociateResolverEndpointIpAddress

Adds IP addresses to an inbound or an outbound Resolver endpoint. If you want to add more than one IP address,

AssociateResolverQueryLogConfig

Associates an Amazon VPC with a specified query logging configuration. Route 53 Resolver logs DNS queries that originate in all of the Amazon VPCs

AssociateResolverRule

Associates a Resolver rule with a VPC. When you associate a rule with a VPC, Resolver forwards all DNS queries

CreateFirewallDomainList

Creates an empty firewall domain list for use in DNS Firewall rules. You can populate the domains for the new list with a file, using ImportFirewallDomains, or with domain strings, using UpdateFirewallDomains.

CreateFirewallRule

Creates a single DNS Firewall rule in the specified rule group, using the specified domain list.

CreateFirewallRuleGroup

Creates an empty DNS Firewall rule group for filtering DNS network traffic in a VPC. You can add rules to the new rule group

CreateResolverEndpoint

Creates a Resolver endpoint. There are two types of Resolver endpoints, inbound and outbound:

CreateResolverQueryLogConfig

Creates a Resolver query logging configuration, which defines where you want Resolver to save DNS query logs that originate in your VPCs.

CreateResolverRule

For DNS queries that originate in your VPCs, specifies which Resolver endpoint the queries pass through,

DeleteFirewallDomainList

Deletes the specified domain list.

DeleteFirewallRule

Deletes the specified firewall rule.

DeleteFirewallRuleGroup

Deletes the specified firewall rule group.

DeleteResolverEndpoint

Deletes a Resolver endpoint. The effect of deleting a Resolver endpoint depends on whether it's an inbound or an outbound

DeleteResolverQueryLogConfig

Deletes a query logging configuration. When you delete a configuration, Resolver stops logging DNS queries for all of the Amazon VPCs that are

DeleteResolverRule

Deletes a Resolver rule. Before you can delete a Resolver rule, you must disassociate it from all the VPCs that you

DisassociateFirewallRuleGroup

Disassociates a FirewallRuleGroup from a VPC, to remove DNS filtering from the VPC.

DisassociateResolverEndpointIpAddress

Removes IP addresses from an inbound or an outbound Resolver endpoint. If you want to remove more than one IP address,

DisassociateResolverQueryLogConfig

Disassociates a VPC from a query logging configuration.

DisassociateResolverRule

Removes the association between a specified Resolver rule and a specified VPC.

GetFirewallConfig

Retrieves the configuration of the firewall behavior provided by DNS Firewall for a

GetFirewallDomainList

Retrieves the specified firewall domain list.

GetFirewallRuleGroup

Retrieves the specified firewall rule group.

GetFirewallRuleGroupAssociation

Retrieves a firewall rule group association, which enables DNS filtering for a VPC with one rule group. A VPC can have more than one firewall rule group association, and a rule group can be associated with more than one VPC.

GetFirewallRuleGroupPolicy

Returns the Identity and Access Management (Amazon Web Services IAM) policy for sharing the

GetResolverDnssecConfig

Gets DNSSEC validation information for a specified resource.

GetResolverEndpoint

Gets information about a specified Resolver endpoint, such as whether it's an inbound or an outbound Resolver endpoint, and the

GetResolverQueryLogConfig

Gets information about a specified Resolver query logging configuration, such as the number of VPCs that the configuration

GetResolverQueryLogConfigAssociation

Gets information about a specified association between a Resolver query logging configuration and an Amazon VPC. When you associate a VPC

GetResolverQueryLogConfigPolicy

Gets information about a query logging policy. A query logging policy specifies the Resolver query logging

GetResolverRule

Gets information about a specified Resolver rule, such as the domain name that the rule forwards DNS queries for and the ID of the

GetResolverRuleAssociation

Gets information about an association between a specified Resolver rule and a VPC. You associate a Resolver rule and a VPC using

GetResolverRulePolicy

Gets information about the Resolver rule policy for a specified rule. A Resolver rule policy includes the rule that you want to share

ImportFirewallDomains

Imports domain names from a file into a domain list, for use in a DNS firewall rule group.

ListFirewallConfigs

Retrieves the firewall configurations that you have defined. DNS Firewall uses the configurations to manage firewall behavior for your VPCs.

ListFirewallDomainLists

Retrieves the firewall domain lists that you have defined. For each firewall domain list, you can retrieve the domains that are defined for a list by calling ListFirewallDomains.

ListFirewallDomains

Retrieves the domains that you have defined for the specified firewall domain list.

ListFirewallRuleGroupAssociations

Retrieves the firewall rule group associations that you have defined. Each association enables DNS filtering for a VPC with one rule group.

ListFirewallRuleGroups

Retrieves the minimal high-level information for the rule groups that you have defined.

ListFirewallRules

Retrieves the firewall rules that you have defined for the specified firewall rule group. DNS Firewall uses the rules in a rule group to filter DNS network traffic for a VPC.

ListResolverDnssecConfigs

Lists the configurations for DNSSEC validation that are associated with the current Amazon Web Services account.

ListResolverEndpointIpAddresses

Gets the IP addresses for a specified Resolver endpoint.

ListResolverEndpoints

Lists all the Resolver endpoints that were created using the current Amazon Web Services account.

ListResolverQueryLogConfigAssociations

Lists information about associations between Amazon VPCs and query logging configurations.

ListResolverQueryLogConfigs

Lists information about the specified query logging configurations. Each configuration defines where you want Resolver to save

ListResolverRuleAssociations

Lists the associations that were created between Resolver rules and VPCs using the current Amazon Web Services account.

ListResolverRules

Lists the Resolver rules that were created using the current Amazon Web Services account.

ListTagsForResource

Lists the tags that you associated with the specified resource.

PutFirewallRuleGroupPolicy

Attaches an Identity and Access Management (Amazon Web Services IAM) policy for sharing the rule

PutResolverQueryLogConfigPolicy

Specifies an Amazon Web Services account that you want to share a query logging configuration with, the query logging configuration that you want to share,

PutResolverRulePolicy

Specifies an Amazon Web Services rule that you want to share with another account, the account that you want to share the rule with,

TagResource

Adds one or more tags to a specified resource.

UntagResource

Removes one or more tags from a specified resource.

UpdateFirewallConfig

Updates the configuration of the firewall behavior provided by DNS Firewall for a single

UpdateFirewallDomains

Updates the firewall domain list from an array of domain specifications.

UpdateFirewallRule

Updates the specified firewall rule.

UpdateFirewallRuleGroupAssociation

Changes the association of a FirewallRuleGroup with a VPC. The association enables DNS filtering for the VPC.

UpdateResolverDnssecConfig

Updates an existing DNSSEC validation configuration. If there is no existing DNSSEC validation configuration, one is created.

UpdateResolverEndpoint

Updates the name of an inbound or an outbound Resolver endpoint.

UpdateResolverRule

Updates settings for a specified Resolver rule. ResolverRuleId is required, and all other parameters are optional.